refactor: 별점 변수 네이밍 수정
